Better Filesystem Capability Checks (e.g. volname & st_birthtime). PR#5 by CoordSpace to original repository by x1ppy https://github.com/x1ppy/freezetag/pull/5
Original message:

When running freezetag mount on non-Mac filesystems, the current set of file
attributes and fuse mount settings cause Runtime Errors with fuse such as fuse:
 unknown option "volname=Music-freezetag" and trigger the persistent process
to instantly crash out.

This patch swaps out checks for OS(es) that don't support these features
(e.g. if platform.system() != 'Windows') for checks that see if freezetag is
running on the one OS variant that does support them: Darwin. This will allow
linux (and maybe windows) users to use freezetag mount with filesystems that
do not support that same filesystem feature set as MacOS.

So far this patch has been tested on the latest versions of Ubuntu Server
and Debian and it clears the hurdles that prevented the new filesystem from
being created, allowing well-behaved FLAC and MP3 files to work as expected.