fix for mount of the  directory that contains two tracks with the same audio data Author: joshbarrass https://github.com/x1ppy/freezetag/pull/7
reezetag seems to support tracks with identical audio but different metadata when
thawing (each checksum is mapped to an array of files, and will then restore the
correct metadata for each). However, when using the mount command, only the one
file is restored properly; all subsequent files will be given this file's metadata.